genAi/requirements.txt
fastapi==0.115.6
pydantic==2.10.2
pre-commit==3.6.2
black==24.2.0
uvicorn[standard]==0.27.1
langchain==0.3.25
weaviate-client
python-dotenv
langchain-openai
langchain-weaviate
langchain-community
langchain-text-splitters
langchain-cohere